Move api conftest fixtures to main conftest file
Want to make sure the Solr-data-generation fixtures created for the api
app are available to tests for other apps, so I'm removing the
api-app-specific conftest file and moving all of those fixtures into the
main conftest file.
