+++ django/sierra/conftest.py
@@ -6,15 +6,18 @@ import pytest
 import importlib
 import redis
 import pysolr
-import datetime
 import pytz
+import hashlib
+from datetime import datetime
 from collections import OrderedDict
 
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.db.models import ObjectDoesNotExist
+from rest_framework import test as drftest
 
-from utils.test_helpers import fixture_factories as ff
+from utils.test_helpers import (fixture_factories as ff,
+                                solr_test_profiles as tp)
 from export import models as em
 from base import models as bm
 from api.models import APIUser
@@ -176,6 +179,48 @@ def global_solr_data_assembler():
         yield make
 
 
+@pytest.fixture(scope='module')
+def solr_profile_definitions(global_solr_conn):
+    """
+    Pytest fixture that returns definitions for Solr profiles, for
+    generating test data via the *_solr_data_factory fixtures.
+    """
+    hs_conn = global_solr_conn('haystack')
+    bib_conn = global_solr_conn('bibdata')
+    return {
+        'location': {
+            'conn': hs_conn,
+            'user_fields': tp.CODE_FIELDS,
+            'field_gens': tp.LOCATION_GENS
+        },
+        'itype': {
+            'conn': hs_conn,
+            'user_fields': tp.CODE_FIELDS,
+            'field_gens': tp.ITYPE_GENS
+        },
+        'itemstatus': {
+            'conn': hs_conn,
+            'user_fields': tp.CODE_FIELDS,
+            'field_gens': tp.ITEMSTATUS_GENS
+        },
+        'item': {
+            'conn': hs_conn,
+            'user_fields': tp.ITEM_FIELDS,
+            'field_gens': tp.ITEM_GENS
+        },
+        'eresource': {
+            'conn': hs_conn,
+            'user_fields': tp.ERES_FIELDS,
+            'field_gens': tp.ERES_GENS
+        },
+        'bib': {
+            'conn': bib_conn,
+            'user_fields': tp.BIB_FIELDS,
+            'field_gens': tp.BIB_GENS
+        }
+    }
+
+
 @pytest.fixture(scope='function')
 def solr_conn():
     """
@@ -326,7 +371,7 @@ def new_export_instance(export_type, export_filter, status,
             export_filter=export_filter(ef_code),
             errors=0,
             warnings=0,
-            timestamp=datetime.datetime.now(pytz.utc)
+            timestamp=datetime.now(pytz.utc)
         )
     return _new_export_instance
 
@@ -566,6 +611,90 @@ def assert_deleted_records_are_not_indexed(assert_records_are_not_indexed):
 
 
 # API App related fixtures
+
+@pytest.fixture(scope='function')
+def api_data_assembler(solr_data_assembler, solr_profile_definitions):
+    """
+    Function-scoped pytest fixture that returns a Solr test data
+    assembler. Records created via this fixture within a test function
+    are deleted when the test function finishes. (For more info about
+    using Solr data assemblers, see the SolrTestDataAssemblerFactory
+    class in utils.test_helpers.fixture_factories.)
+    """
+    return solr_data_assembler(tp.SOLR_TYPES, tp.GLOBAL_UNIQUE_FIELDS,
+                               tp.GENS, solr_profile_definitions)
+
+
+@pytest.fixture(scope='module')
+def global_api_data_assembler(global_solr_data_assembler,
+                              solr_profile_definitions):
+    """
+    Module-scoped pytest fixture that returns a Solr test data
+    assembler. Records created via this fixture persist while all tests
+    in the module run. (For more info about using Solr data assemblers,
+    see the SolrTestDataAssemblerFactory class in
+    utils.test_helpers.fixture_factories.)
+    """
+    return global_solr_data_assembler(tp.SOLR_TYPES, tp.GLOBAL_UNIQUE_FIELDS,
+                                      tp.GENS, solr_profile_definitions)
+
+
+@pytest.fixture(scope='module')
+def api_solr_env(global_api_data_assembler):
+    """
+    Pytest fixture that generates and populates Solr with some random
+    background test data for API integration tests. Fixture is module-
+    scoped, so test data is regenerated each time the test module runs,
+    NOT between tests.
+    """
+    assembler = global_api_data_assembler
+    gens = assembler.gen_factory
+    loc_recs = assembler.make('location', 10)
+    itype_recs = assembler.make('itype', 10)
+    status_recs = assembler.make('itemstatus', 10)
+    bib_recs = assembler.make('bib', 100)
+    item_recs = assembler.make('item', 200,
+        location_code=gens.choice([r['code'] for r in loc_recs]),
+        item_type_code=gens.choice([r['code'] for r in itype_recs]),
+        status_code=gens.choice([r['code'] for r in status_recs]),
+        parent_bib_id=gens(tp.choose_and_link_to_parent_bib(bib_recs))
+    )
+    eres_recs = assembler.make('eresource', 25)
+    assembler.save_all()
+    return assembler
+
+
+@pytest.fixture
+def api_client():
+    """
+    Pytest fixture that returns a new rest_framework.test.APIClient
+    object.
+    """
+    return drftest.APIClient()
+
+
+@pytest.fixture
+def simple_sig_auth_credentials():
+    """
+    Pytest fixture that generates auth headers for the given `api_user`
+    instance and optional `request_body` string so that a request using
+    the custom api.simpleauth.SimpleSignatureAuthentication mechanism
+    authenticates.
+    """
+    def _simple_sig_auth_credentials(api_user, request_body=''):
+        since_1970 = (datetime.now() - datetime(1970, 1, 1))
+        timestamp = str(int(since_1970.total_seconds() * 1000))
+        hasher = hashlib.sha256('{}{}{}{}'.format(api_user.user.username,
+                                                  api_user.secret, timestamp,
+                                                  request_body))
+        signature = hasher.hexdigest()
+        return {
+            'HTTP_X_USERNAME': 'test',
+            'HTTP_X_TIMESTAMP': timestamp,
+            'HTTP_AUTHORIZATION': 'Basic {}'.format(signature)
+        }
+    return _simple_sig_auth_credentials
+
 
 @pytest.fixture(scope='function')
 def apiuser_with_custom_defaults():
